Control.Delimited

Description

Delimited continuations featuring answer-type polymorphism, via parameterized monads.

There is a tutorial available in Control.Delimited.Tutorial.

Delimited continuations

data Delim s t b

The type of a delimited continuation, which is answer-type polymorphic.

Functions of type a -> Delim s t b can be thought of as functions of type a / s -> b / t, which means given an a we return a b, changing the answer type of the continuation from s to t.

If a Delim does not capture the computation using one of the various shift operators (or shift does not change the answer type,) then the term is polymorphic in the answer type.

>>> :t runDelim $ reset $ shift2 ret !>>= \r -> ret (r + (1 :: Int))
runDelim $ reset $ shift2 ret !>>= \r -> ret (r + (1 :: Int))
  :: Int -> forall a'. Delim a' a' Int

Note how the quantified variable a' is both the input and output answer type: thus, it cannot change from what is (in this case, the answer type of the enclosing reset is Int.) We use shift2 here, which uses rank-2 typing to ensure the type variable does not 'escape', much like the ST monad.

reset :: Delim s t s -> Delim s' s' t

Delimit a computation.

A family of shift operators

shift0 :: ((b -> s) -> t) -> Delim s t b

Clear the current continuation and invoke our handler with it bound as a parameter.

This is the most pure definition of shift.

shift1 :: ((b -> s) -> Delim a t a) -> Delim s t b

Clear the current continuation and invoke our handler with it bound as a parameter.

This is a simple Haskell98 definition of shift that does not enforce true answer type polymorphism by abstracting over the internal s and t type variables.

shift2 :: ((b -> forall a'. Delim a' a' s) -> Delim a t a) -> Delim s t b

Clear the current continuation and invoke our handler with it bound as a paramter.

This definition of shift uses Rank-2 types to ensure the answer type is in fact polymorphic: note the type of the captured continuation is of type Delim t' t' s.

Executing delimited computations

runDelim :: Delim t t t -> t

Run a delimited computation.

Re-exports for convenience

module Control.Indexed.Monad