Control.Indexed.Monad

Description

Indexed Monads.

When using RebindableSyntax with GHC, it's possible to overload regular do-notation to use the IxMonad typeclass. See the section "Using do-notation" in the tutorial module Control.Delimited.Tutorial.

Currently, indexed transformers are not provided. In the future, this module may be replaced by the indexed or index-core libraries which provide more general types while still having do-notation, and providing transformers. In light of that, this module will be expanded as needed.

Indexed Monads

class IxApplicative m => IxMonad m where

Indexed Monads.

Regular monads can be lifted into this type class using lift.

Methods

(!>>=) :: m t u a -> (a -> m s t b) -> m s u b

Indexed >>=.

joinI :: m t u (m s t a) -> m s u a

Indexed join.

failI :: String -> m s s a

Indexed fail.

Lifting ordinary Monads

data MW m s t a

This type lifts any regular Monad into an IxMonad.

lift :: Monad m => m a -> MW m s t a

This method lifts a regular Monad into MW which is an instance of IxMonad.

runI :: Monad m => MW m s t a -> m a

This demotes a indexed monad into a regular monad. Useful for when you're using e.g. RebindableSyntax and want to do IO.

Operators

(!>>) :: IxMonad m => m s t a -> m t u b -> m s u b

Indexed >>.

(!>=>) :: IxMonad m => (a -> m t u b) -> (b -> m s t c) -> a -> m s u c

Left-to-right Kleisli composition of indexed Monads.

(<=<!) :: IxMonad m => (b -> m s t c) -> (a -> m t u b) -> a -> m s u c

Right-to-left Kleisli composition of indexed Monads.

(=<<!) :: IxMonad m => (a -> m s t b) -> m t u a -> m s u b

Right-to-left version of !>>=.

Other functions

liftIxM :: IxMonad m => (a -> b) -> m s t a -> m s t b

This is liftM for indexed Monads.

liftIxM2 :: IxMonad m => (a -> b -> c) -> m t u a -> m s t b -> m s u c

This is liftM2 for indexed Monads.