Control.Delimited.Tutorial

Description

This module provides an introductory tutorial in the "Introduction" section and beyond, followed by examples, and by some discussion behind the theory of the library with references.

Introduction

asai is a minimal library for delimited continuations, a slice of a continuation that can be invoked as a function and composed.

An primer on continuations

Continuations are a well known abstraction for 'picking up where you left off.' When you use call/cc traditionally in something like scheme, you pass it a function f which receives a function k, and when k is invoked you 'call the continuation' and return where call/cc left off. In this sense, k is a reification of the state you were in when you invoked f.

But call/cc is, as some say, 'overrated.' For one, it is not possible to return a value from the continuation; you merely pick up where you left off. By fixing this we get delimited continuations: an abstraction that allows us to slice up a continuation and compose them. But at this point, we suffer in a statically typed language: the type of shift and reset are not polymorphic in their return.

In the same way that continuations form a monad, so do delimited continuations. This package provides a delimited continuation monad which implements truly polymorphic shift and reset operators, via answer type polymorphism, which allows the results of a delimited computation to vary.

This implementation (using indexed monads) was first implemented by - and this package derived from - Oleg Kiselyov [1]. It directly implements the typing rules of Kenichi Asai's lambda/shift calculus [2], featuring answer type polymorphism and modification.

A general tutorial on delimited continuations in OchaCaml (with code in Haskell and OCaml) from Asai/Kiselyov is available [3]. A more traditional delimited continuation monad is available in the CC-delcont package [4]. The CC-delcont tutorial [5] serves as a basis for this tutorial (thanks to Dan Doel!)

Undelimited

Lorem ipsum...

Delimited

Lorem ipsum...

>>> runDelim $ reset $ shift1 (\_ -> ret "hello") !>>= \r -> ret (r + 1)
"hello"

Examples

In the subsequent sections, we'll cover some examples of using this library in a few simple forms, which should form a basis for reasoning about how to use it. The same ideas should apply to other delimited continuation libraries or interfaces (i.e. scheme, ocaml, ochacaml, etc.)

Simple answer-type modification

Lorem ipsum...

>>> runDelim $ reset $ shift1 (\_ -> return "hello") >>= \r -> return (r + 1)
"hello"

First-class printf

Lorem ipsum...

Walking binary trees

Here, we will...

 
 -- Walking trees with coroutines.
 -- Based on:
 --    <http://okmij.org/ftp/continuations/ContExample.hs>
 module Tree where

 import Control.Delimited

 -- Binary trees
 data Tree a
   = Leaf
   | Node (Tree a) (Tree a) a
   deriving (Show, Eq)

 make_tree :: Int -> Tree Int
 make_tree j = go j 1
   where go 0 _ = Leaf
         go i x = Node (go (i-1) $ 2*x) (go (i-1) $ 2*x+1) x

 tree1 = make_tree 3
 tree2 = make_tree 4

 -- Coroutines as lazy monadic lists.
 data Coro a
   = Done
   | Resume a (forall s. Delim s s (Coro a))

 walk_tree x = runDelim (walk_tree' x !>> ret Done)

 walk_tree' Leaf = ret ()
 walk_tree' (Node l r x) =
   walk_tree' l !>>
   yield x      !>>
   walk_tree' r
   where
     yield n = shift2 (\k -> ret $ Resume n $ k ())

 walk1 t = go (walk_tree t)
   where
     go Done         = return ()
     go (Resume x k) = print x >> go (runDelim k)

The full code is available in examples/Tree.hs.

Other notes

Here we discuss some of the design aspects of the library, particularly for those wondering why we need indexed (or indexed) monads, and how we can reconcile this with do-notation.

Indexed monads

While reading this tutorial, you may wonder why we use special operators (like !>>= that mimmick >>=) for delimited computations, instead of regular operators from the Monad typeclass. The reason for this is that in order to ensure the answer type of delimited computation is polymorphic, we need the monad to 'carry' the answer types around.

Consider the vanilla Monad typeclass. It is defined like this (with explicit kind signatures):

 class Monad (m :: * -> *) where ...

The m type constructor abstracts over a single type variable. It is possible to make types with multiple type variables an instance of Monad of course, but their non-abstracted type variables must be fixed. As an example, considering the instance for Either e:

 instance Monad (Either e) where ...

Note the type variable e is fixed over the definition of a term of type Either e :: * -> *. If you have something like:

 thing :: a -> Either String a
 thing a = do
   ...

Then in the body we may say:

 x <- Left "oh no!"

But we can never say:

 x <- Left False

because e is fixed to String.

Indexed monads solve this problem by 'expanding' the kind of m in the Monad typeclass. The result is IxMonad, which is defined as:

 class IxMonad (m :: * -> * -> * -> *) where
   ret    :: t -> m a a t
   (!>>=) :: m b g s -> (s -> m a b t) -> m a g t

Note the new type variables: these represent the input and output answer types of a delimited computation. We can see that ret is fully polymorphic in its answer type: a statement of ret foo in a Delim simply does not change the answer type at all. Note the answer types present in !>>=: we have an answer type of a b and b g, meaning we can get an overall answer type of a g.

This polymorphism in the definition of !>>= is what gives us answer type polymorphism: it means delimited computations may change the output answer type.

Using do-notation

It's possible to use GHC's RebindableSyntax extension to re-define do notation to use the IxMonad type class.

Begin your module by hiding the regular Monad methods, and then redefine >>= and return. Feel free to redefine other operators too. Here's an example (you'll need to fix the LANGUAGE pragma yourself on the first line, since Haddock eats it otherwise):

 [-# LANGUAGE RebindableSyntax, NoImplicitPrelude #-]
 module Foo where

 import Control.Indexed.Prelude
 import Control.Delimited

 -- Now use 'do' notation instead of the indexed bind/return
 -- functions.

 -- You can lift regular monads into indexed monads using
 -- 'lift' and 'runI'
 io1 :: IO ()
 io1 = runI $ do
   lift $ putStrLn "hi!"
   lift $ putStrLn "hi!"
   return ()

 test1 :: String
 test1 = runDelim $ reset $ do
   r <- shift1 (\_ -> return "hello")
   return (r + 1)
 -- This is equivalent to the OchaCaml term:
 --   reset (fun () -> 1 + shift (fun _ -> "hello")) ;;

See examples/Simple.hs (included in the distribution) for several more examples.

Rank-2 typing

This package requires GHC's RankNTypes extension, as it uses a rank-2 type for the definition of the shift operators shift2 and shift3. The original implementation by Kiselyov is Haskell98 (he defines shift as the shift1 provided in this package, as opposed to shift2 which matches the typing rules of the lambda/shift calculus.)

Strictly speaking, the rank-2 type is probably not necessary, but it is not very controversial either, and it makes the intent much clearer.

You do not necessarily need to enable RankNTypes to use this package, but in certain cases where you need to be polymorphic over recursive uses of Delim for example, it will be necessary to quantify the answer type variables with a rank-2 type. However, because the provided family of shift operators is equivalent, you should be able to get by without RankNTypes should you choose.

Further reading

Lorem ipsum...

References

Genuine shift\reset in Haskell98, by Kiselyov, on haskell-cafe: http://okmij.org/ftp/continuations/implementations.html#genuine-shift

Polymorphic Delimited Continuations, by Asai, Kameyama in APLAS '07: http://logic.cs.tsukuba.ac.jp/~kam/paper/aplas07.pdf

Introduction to programming with shift and reset, by Kiselyov, Asai, in CW2011: http://okmij.org/ftp/continuations/index.html#tutorial

CC-delcont: Delimited continuations and dynamically scoped variables: http://hackage.haskell.org/package/CC-delcont

CC-delcont introduction: <http:www.haskell.orghaskellwikiLibraryCC-delcont#CC-delcont>